Port over to modular folder structure from layered
