+++ src/index.ts
@@ -1,3 +1,11 @@
-export * from './client/GithubClient'
-export * from './errors/errors'
-export * from './types/index.types'
+export { GithubClient } from './client/GithubClient';
+
+export * from './modules/commits/commit.types';
+export * from './modules/pull-requests/pull-request.types';
+export * from './modules/repositories/repository.types';
+export * from './modules/users/user.types';
+
+export * from './shared/errors/GithubSdkError';
+export * from './shared/errors/GithubApiError';
+export * from './shared/errors/InvalidTokenError';
+export * from './shared/errors/MissingConfigError';
