+++ src/modules/commits/commit.dto.ts
@@ -1,0 +1,44 @@
+import { UserDTO } from "../users/user.dto";
+
+interface CommitAuthorDTO {
+    name: string;
+    email: string;
+    date: string;
+}
+
+interface CommitTreeDTO {
+    sha: string;
+    url: string;
+}
+
+interface VerificationDTO {
+    verified: boolean;
+    reason: string;
+    signature: string | null;
+    payload: string | null;
+}
+
+interface CommitDetailsDTO {
+    author: CommitAuthorDTO;
+    committer: CommitAuthorDTO;
+    message: string;
+    tree: CommitTreeDTO;
+    url: string;
+    comment_count: number;
+    verification: VerificationDTO;
+}
+
+export interface CommitDTO {
+    sha: string;
+    node_id: string;
+    commit: CommitDetailsDTO;
+    url: string;
+    html_url: string;
+    author: UserDTO | null;
+    committer: UserDTO | null;
+    parents: {
+        sha: string;
+        url: string;
+        html_url: string;
+    }[];
+}
