+++ src/shared/errors/InvalidTokenError.ts
@@ -1,0 +1,8 @@
+import { GithubSdkError } from "./GithubSdkError";
+
+export class InvalidTokenError extends GithubSdkError {
+    constructor() {
+        super('Invalid GitHub Token');
+        this.name = 'InvalidTokenError';
+    }
+}
