--- src/types/index.types.ts
@@ -1,3 +1,0 @@
-export * from './pullrequest.types';
-export * from './user.types';
-export * from './commit.types';
